[CMake] Add LINK_LIBS and LINK_COMPONENTS options
This patch adds CMake options to add_lldb_library and add_lldb_executable for specifying LLVM components and direct library links.

This patch is NFC, but it is a small separable bit of a series of much larger patches that I'll be landing over the next day or two.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@293647 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 31b5564935eed51bfc98ca86ae4ae3632b47af3d
apple-llvm-split-dir: lldb/