[CMake] Add explicit dependencies to plugins
Summary:
This patch does two things. First it updates all the ABI plugins with accurate dependencies, and second it adds a tracking mechanism for add_lldb_library to denote plugin libraries, allowing us to build up a list of all the configured plugins.

This list of generated plugins will be used during generating liblldb so that we can link all the plugins into the library.

If this patch looks good I will update all the other plugins in subsequent patches.

Reviewers: labath, zturner

Subscribers: nemanjai, mgorny, lldb-commits, jgosnell

Differential Revision: https://reviews.llvm.org/D29348

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@293696 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 2b20d05c64f08068fa1893448c7da8c2c385f13b
apple-llvm-split-dir: lldb/