[ARM/AArch ISel] SwiftCC: First parameters that are marked swiftself are not 'this returns'
We mark X0 as preserved by a call that passes the returned parameter.

 x0 = ...
 fun(x0) // no implicit def of x0

This no longer is valid if we pass the parameter in a different register then
the returned value as is the case with a swiftself parameter (passed in x20).

x20 = ...
fun(x20) // there should be an implict def of x8

rdar://30425845

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@294527 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 7296f4cb830c7e425f2b92b5e34ff6542931fd7c)

apple-llvm-split-commit: 30e7df2afbc872095ec71fc56707e1bc41c1f739
apple-llvm-split-dir: llvm/