[CodeGen][ObjC] Use the type of the captured field of the enclosing block or lambda.
This is a follow-up to r281682, which fixed a bug in computeBlockInfo
where the captured VarDecl's type, rather than the captured field type
of the enclosing lambda or block, was used to compute the layout of a
block.

This commit makes similar changes to enterBlockScope. This is necessary
to correctly determine whether a block capture requires cleanup.

rdar://problem/30388124

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@295034 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit c644cb6f1e7f85c9b918b6d4917dd161aa207960)

apple-llvm-split-commit: b21a7266b78524fc78a10ab425555eb08e6afc4b
apple-llvm-split-dir: clang/