Changed builld-llvm.py to use .json files
LLDB has many branches in a variety of repositories.
The build-script.py file is subtly different for each set.
This is unnecessary and causes merge headaches.

This patch makes build-llvm.py consult a directory full
of .json files, each one of which matches a particular
branch using a regular expression.

This update to the patch introduces a FALLBACK file
whose contents take precedence if the current branch
could not be identified.  If the current branch could be
identified, FALLBACK is updated, allowing the user to
e.g. cut branches off of known branches and still have
the automatic checkout mechanism work.

It also documents all of this.

This patch adapts the mechanism to work in github.

Differential revision: https://reviews.llvm.org/D30275

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@295922 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 90a0aa1064d45f5390f2d73a5269605623362df9
apple-llvm-split-dir: lldb/