[InlineCost] Fix codesize issue caused by r286814
This modifies InlineCost.cpp and several tests in order to fix code size issues
caused by r286814.

Upstream, inline cost calculations remove the overhead of an instruction and the
penalty of a call at each call-site. This causes too many things to be inlined
in some projects, so it has been removed.

Inlining thresholds in the test have been modified to reflect this.

rdar://30870912

apple-llvm-split-commit: af2d0351e2b108a9caee7911998ba92454a8ea63
apple-llvm-split-dir: llvm/