[DAGCombine] Support FMF contract in fused multiple-and-sub too
This is a follow-on to r299096 which added support for fmadd.

Subtract does not have the case where with two multiply operands we commute in
order to fuse with the multiply with the fewer uses.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@299572 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit c0a481413e8c6c18e7744041f5a7f26674f9b4ea)

apple-llvm-split-commit: 2bbdfc6d2839f4355e4d7b068885f71abc952e9a
apple-llvm-split-dir: llvm/