[Modules] Improve redefinition errors pointing to the same header
Diagnostics related to redefinition errors that point to the same header file
do not provide much information that helps fixing the issue. In modules context
it usually happens because of a non modular include, in non-module context it
might happen because of the lack of header guardas.

Enhance the diagnostics for this particular scenario.

rdar://problem/30140314

apple-llvm-split-commit: a664f83fc5818d40bce795c9816a0f3c0b92d6f6
apple-llvm-split-dir: clang/