Add a fix-it for -Wunguarded-availability
This patch adds a fix-it for the -Wunguarded-availability warning. This fix-it
is similar to the Swift one: it suggests that you wrap the statement in an
`if (@available)` check. The produced fixits are indented (just like the Swift
ones) to make them look nice in Xcode's fix-it preview.

rdar://31680358

Differential Revision: https://reviews.llvm.org/D32424

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@302253 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 90500c1350ce925c24aa3da49b42657140999764)

apple-llvm-split-commit: 4cc87092bc0d7bad7c1a90d0de19ee039e5e5bd8
apple-llvm-split-dir: clang/