Disable runtime enforcement of exclusivity during expression evaluation.
The debugger needs to be able to read the value of variables regardless of
whether they're marked as exclusively owned/locked by the swift runtime. This
patch modifies SwiftUserExpressions to poke the swift runtime before running
an expression to disable exclusivity enforcement, and again after the expression
has finished to re-enable enforcement.

<rdar://problem/31715465>

apple-llvm-split-commit: 235dda97a00cdc9e4bbb7dfe258c36cd5afff589
apple-llvm-split-dir: lldb/