[X86][AVX-512] Don't raise inexact in ceil, floor, round, trunc.
The non-AVX-512 behavior was changed in r248266 to match N1778
(C bindings for IEEE-754 (2008)), which defined the four functions
to not raise the inexact exception ("rint" is still defined as raising
it).

Update the AVX-512 lowering of these functions to match that: it should
not be different.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@306299 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 699f4c431de74c28e66ace05d9e83223993c7a13)

rdar://30593055

apple-llvm-split-commit: e6869aee627ad70da797fea037d86ce9c56f004b
apple-llvm-split-dir: llvm/