[coverage] Special-case calls to noreturn functions.
The code after a noreturn call doesn't execute.

The pattern in the testcase is pretty common in LLVM (a switch with
a default case that calls llvm_unreachable).

The original version of this patch was reverted in r309995 due to a
crash. This version includes a fix for that crash (testcase in
test/CoverageMapping/md.cpp).

Differential Revision: https://reviews.llvm.org/D36250

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@310406 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 2cac4c7ed196ae69977c4624745e246b2ae8eedf)

 Conflicts:
	test/CoverageMapping/switch.cpp

apple-llvm-split-commit: b9c140647f55f22aad0a62c5b3f9694d9d124d88
apple-llvm-split-dir: clang/