Extract IRGen's constant-emitter into its own helper class and clean up the interface.
The ultimate goal here is to make it easier to do some more interesting
things in constant emission, like emit constant initializers that have
ignorable side-effects, or doing the majority of an initialization
in-place and then patching up the last few things with calls.  But for
now this is mostly just a refactoring.

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@310964 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 5ac85c20ddfde03a8b1453971adf20b93bd67b3d
apple-llvm-split-dir: clang/