[diagtool] Change default tree behavior to print only flags
This patch changes the default behavior of `diagtool tree` to only
display warning flags and not the internal warnings flags. The latter is
an implementation detail of the compiler and usually not what the users
wants.

Furthermore, flags that are enabled by default are now also printed in
green. Originally, this was only the case for the diagnostic names.

Differential revision: https://reviews.llvm.org/D37390

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@312546 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 9a30feb06ef1ecc59dcd9dd3dfb46478c476be1d)

apple-llvm-split-commit: 849a311b40a534f2d944c2ff34f09a900f21989d
apple-llvm-split-dir: clang/