Work around a bug in the C++ expression parser.
When the expression parser does name resolution for local
variables in C++ closures it doesn't give the local name
priority over other global symbols of the same name.  heap.py
uses "info" which is a fairly common name, and so the commands
in it fail.  This is a workaround, just use lldb_info not info.

<rdar://problem/34026140>


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@314959 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 3518796365f08b913a327f9395a4a8fa9f6f293a
apple-llvm-split-dir: lldb/